@@ -18,17 +18,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.0.0",
+    "@angular/animations": "^18.2.4",
     "@angular/cdk": "^17.3.10",
-    "@angular/common": "^17.0.0",
-    "@angular/compiler": "^17.0.0",
-    "@angular/core": "^17.0.0",
-    "@angular/forms": "^17.0.0",
-    "@angular/platform-browser": "^17.0.0",
-    "@angular/platform-browser-dynamic": "^17.0.0",
-    "@angular/platform-server": "^17.0.0",
-    "@angular/router": "^17.0.0",
-    "@angular/ssr": "^17.0.6",
+    "@angular/common": "^18.2.4",
+    "@angular/compiler": "^18.2.4",
+    "@angular/core": "^18.2.4",
+    "@angular/forms": "^18.2.4",
+    "@angular/platform-browser": "^18.2.4",
+    "@angular/platform-browser-dynamic": "^18.2.4",
+    "@angular/platform-server": "^18.2.4",
+    "@angular/router": "^18.2.4",
+    "@angular/ssr": "^18.2.4",
     "@tinkoff/ng-polymorpheus": "^4.3.0",
     "bootstrap": "^4.6.2",
     "express": "^4.18.2",
@@ -37,16 +37,16 @@
     "marked": "14.1.2",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.14.2"
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
     "@angular-builders/custom-webpack": "^17.0.2",
-    "@angular-devkit/build-angular": "^17.0.6",
+    "@angular-devkit/build-angular": "^18.2.4",
     "@angular-eslint/eslint-plugin": "17.4.0",
     "@angular-eslint/eslint-plugin-template": "17.4.0",
     "@angular-eslint/template-parser": "17.4.0",
-    "@angular/cli": "^17.0.6",
-    "@angular/compiler-cli": "^17.0.0",
+    "@angular/cli": "^18.2.4",
+    "@angular/compiler-cli": "^18.2.4",
     "@types/express": "^4.17.21",
     "@types/jasmine": "~5.1.0",
     "@types/node": "^18.18.0",
@@ -68,6 +68,6 @@
     "karma-jasmine-html-reporter": "~2.1.0",
     "remark-loader": "^6.0.0",
     "ts-node": "^10.9.2",
-    "typescript": "~5.2.2"
+    "typescript": "~5.5.4"
   }
 }
