@@ -1,19 +1,33 @@
-import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
-import { MenuItemModel } from "../../../../models";
-import { Observable } from "rxjs";
-import { MenuStateService } from "./services";
+import {
+  ChangeDetectionStrategy,
+  Component,
+  computed,
+  effect,
+  input,
+  Input,
+  Signal,
+} from '@angular/core';
+import { MenuItemModel } from '../../../../models';
+import { Observable } from 'rxjs';
+import { MenuStateService } from './services';
 
 @Component({
   selector: 'app-menu',
   templateUrl: './menu.component.html',
   styleUrl: './menu.component.scss',
-  changeDetection: ChangeDetectionStrategy.OnPush
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class MenuComponent {
-  public displayingSubmenuSet$: Observable<Set<string>> = this.menuStateService.displayingSubmenuSet$;
+  public displayingSubmenuSet: Signal<Set<string>> =
+    this.menuStateService.displayingSubmenuSet;
 
-  @Input()
-  public menu: MenuItemModel[] = [];
+  public displayingSubmenuSetDEBUG: any = computed(() => ({
+    ...this.menuStateService.displayingSubmenuSet(),
+  }));
+
+  public menu = input<MenuItemModel[]>([]);
+  public level = input(0);
+  public nextLevel = computed(() => this.level() + 1);
 
   constructor(private readonly menuStateService: MenuStateService) {}
 
